Add concept of instance
The .boss.yml file originally had a section for platforms; now
there is also a section for 'profiles'. Combining a platform and
profile gives an 'instance'. This is similar to test-kitchen,
which combines platform and 'suite' to make an instance. I didn't
use the term suite because I don't think it fits as well with
Ansible, which does not have a 'run_list' like Chef. The purpose
of the profile will be to add extra-vars to a build so that images
can be customized while still using the same role.