Add schema validation of config with voluptuous
In addition to validating the config, the config is now transformed
so that it is a dictionary with the top level keys being the
'instances' created from all platform-profile pairs. The values of
the keys are dictionaries created from the platform and profile
configs being merged, along with any global settings from the
'driver' config.