Add ability to name AMI
The AMI naming uses a Python formatted string. By default this is:

'%(role)s-%(profile)s-%(version)s-%(platform)s'

The values of the string variables are taken from the configuration
for the instance. The `version` variable comes from the file
meta/main.yml in the Ansible role, so if using this in the AMI
name, then a version attribute must be added to that file.