Add grouping to inventory in preparation for adding test hosts.
Bossimage will soon support creating a host for testing, based on
an AMI previously built for an instance. By adding grouping to the
inventory, the AMI build host can be in one group and the test
instance in another. This commit has no visible effect on the
functionality, but makes it possible to add that change next.